+++ lihil/__init__.py
@@ -1,24 +1,27 @@
+from ididi import AsyncScope as AsyncScope
 from ididi import Graph as Graph
+from ididi import Resolver as Resolver
 
 from .constant import status as status
-from .interface import Json as Json
+from .interface import Body as Body
+from .interface import Empty as Empty
 from .interface import Form as Form
+from .interface import Json as Json
+from .interface import Path as Path
 from .interface import Payload as Payload
+from .interface import Query as Query
 from .interface import Resp as Resp
 from .interface import Stream as Stream
 from .interface import Text as Text
 from .interface import Use as Use
-from .interface import Body as Body
-from .interface import Query as Query
-from .interface import Path as Path
-from .interface import Empty as Empty
 from .lihil import Lihil as Lihil
 from .problems import HTTPException as HTTPException
 from .routing import Route as Route
 from .vendor_types import Request as Request
 from .vendor_types import Response as Response
 from .vendor_types import UploadFile as UploadFile
+
 # from .server.runner import run as run
 
-VERSION = "0.1.10"
+VERSION = "1.5.4"
 __version__ = VERSION