+++ lihil/interface/__init__.py
@@ -1,19 +1,7 @@
-from typing import (
-    Annotated,
-    Any,
-    Callable,
-    Literal,
-    Protocol,
-    Self,
-    TypeGuard,
-    dataclass_transform,
-    get_args,
-)
+from typing import Callable, Literal, TypeGuard, get_args
 
 from msgspec import Struct as Struct
 from msgspec import field as field
-from msgspec.structs import asdict as struct_asdict
-from msgspec.structs import replace as struct_replace
 
 from lihil.interface.asgi import HTTP_METHODS as HTTP_METHODS
 from lihil.interface.asgi import ASGIApp as ASGIApp
@@ -31,8 +19,16 @@ from lihil.interface.marks import Resp as Resp
 from lihil.interface.marks import Stream as Stream
 from lihil.interface.marks import Text as Text
 from lihil.interface.marks import Use as Use
-from lihil.interface.marks import EMPTY_RETURN_MARK
 from lihil.interface.marks import lhl_get_origin as lhl_get_origin
+from lihil.interface.struct import Base as Base
+from lihil.interface.struct import Empty as Empty
+from lihil.interface.struct import IDecoder as IDecoder
+from lihil.interface.struct import IEncoder as IEncoder
+from lihil.interface.struct import ParamBase as ParamBase
+from lihil.interface.struct import Payload as Payload
+from lihil.interface.struct import Record as Record
+from lihil.interface.struct import CustomDecoder as CustomDecoder
+from lihil.interface.struct import CustomEncoder as CustomEncoder
 
 type ParamLocation = Literal["path", "query", "header", "body"]
 type Func[**P, R] = Callable[P, R]
@@ -61,88 +57,3 @@ class _Missed:
 
 
 MISSING = _Missed()
-
-
-class IDecoder[T](Protocol):
-    def __call__(self, content: Any, /) -> T: ...
-
-
-class IEncoder[T](Protocol):
-    def __call__(self, content: T, /) -> bytes: ...
-
-
-# class IProblem[T](Protocol):
-#     """
-#     https://www.rfc-editor.org/rfc/rfc9457.html
-#     """
-
-#     @property
-#     def type_(self) -> str: ...
-#     @property
-#     def title(self) -> str: ...
-#     @property
-#     def status(self) -> int: ...
-#     @property
-#     def detail(self) -> T: ...
-#     @property
-#     def instance(self) -> str: ...
-
-
-class Base(Struct):
-    "Base Model for all internal struct, with Mapping interface implemented"
-
-    def keys(self) -> tuple[str, ...]:
-        return self.__struct_fields__
-
-    def __getitem__(self, key: str) -> Any:
-        return getattr(self, key)
-
-    def asdict(self):
-        return struct_asdict(self)
-
-    def replace(self, /, **changes: Any) -> Self:
-        return struct_replace(self, **changes)
-
-
-@dataclass_transform(kw_only_default=True)
-class KWBase(Base, kw_only=True): ...
-
-
-class ParamBase[T](Base):
-    type_: type
-    decoder: IDecoder[T]
-
-
-@dataclass_transform(frozen_default=True)
-class Record(Base, frozen=True, gc=False, cache_hash=True): ...  # type: ignore
-
-
-@dataclass_transform(frozen_default=True)
-class Payload(Record, frozen=True, gc=False):
-    """
-    a pre-configured struct that is frozen, gc_free, tagged with "typeid"
-    """
-
-
-class CustomEncoder(Base):
-    encode: Callable[[Any], bytes]
-
-
-class CustomDecoder(Base):
-    """
-    class IType: ...
-
-    def decode_itype()
-
-
-    async def create_user(i: Annotated[IType, CustomDecoder(decode_itype)])
-    """
-
-    decode: Callable[[bytes | str], Any]
-
-
-def empty_encoder(param: Any) -> bytes:
-    return b""
-
-
-type Empty = Annotated[Literal[None], CustomEncoder(empty_encoder), EMPTY_RETURN_MARK]