+++ lihil/oas/doc_ui.py
@@ -2,7 +2,7 @@ from typing import Any
 
 from starlette.responses import HTMLResponse
 
-from lihil.problems import DetailBase
+from lihil.interface.problem import DetailBase
 from lihil.utils.phasing import encode_json
 
 problem_ui_default_parameters: dict[str, Any] = {